Move LUT encoder and decoder to the same file.
